[COS] Integration tests can skip cos tests on arm64 (#653)
* Integration tests can skip cos tests on arm64

* Move cos tests to its own module

* Always run cos integration tests
