+++ CONTRIBUTING.md
@@ -201,10 +201,6 @@ Running the integration tests with extra arguments can be accomplished with
 tox run -e integration-tests -- --positional --arguments
 ```
 
-#### Canonical observability integration
-
-The COS integration tests are optional as these are slow/heavy tests. Currently, this suite only runs on LXD. If you are modifying something related to the COS integration, you can validate your changes through integration testing using the flag `--cos`. Also, when submitting a Pull Request with changes related to COS, you must include the `[COS]` tag in your Pull Request description. This will instruct GitHub Actions to execute the respective validation tests against your changes.
-
 #### Useful arguments
 
 `--keep-models`: Doesn't delete the model once the integration tests are finished
