+++ test_requirements.txt
@@ -1,3 +1,4 @@
+async-lru
 juju
 pydantic<2
 pylxd
